+++ RELEASES.md
@@ -1,6 +1,7 @@
 # Unreleased
 ## Changed
 - Bumped MSRV to `1.63.0`
+- Bumped dependencies to `bls12_381 0.9`, `ff 0.14`, `group 0.14`.
 
 # 0.10.0
 ## Changed